requirements.txt
langchain
pandas
numpy
statsmodels
python-dotenv
langchain-openai
openai
streamlit
langchain_core
